+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Render Texture as direct video screen targets. The core will blit the screen to these RTs every frame when the video is playing.
+- Fix aspect ratio button in video screen configurator.
+
+### Changed
+- Canvas Utility now able to handle multiple hierarchy of non-uniform scale adjustment, making sure the UI don't stretched.
+- Updated exhibition setup to use new active region auto play component.
 
 ## 1.7.0-beta.1 - 2026-04-13
 ### Added